server-crawler/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.12.0
python-dateutil==2.8.2
aiohttp==3.9.3
httpx==0.26.0 